reorganised helper functions
